requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python>=4.10.0
pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
numpy>=1.24.0
